# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Groundwork Engineering Evidence API",
    "version": "3.0.0",
    "description": "Workspace-scoped evidence, change analysis, connector, policy, release, document, and grounded retrieval API."
  },
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "WorkspaceId": { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "ConnectionId": { "name": "connectionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "ChangeId": { "name": "changeId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
    },
    "schemas": {
...
      "ConnectionRequest": {
        "type": "object",
        "required": ["provider", "externalAccountId", "displayName"],
        "properties": {
          "provider": { "type": "string", "enum": ["GITHUB", "JIRA", "CONFLUENCE", "MANUAL", "DEMO"] },
          "externalAccountId": { "type": "string" },
          "displayName": { "type": "string" },
          "scopes": { "type": "array", "items": { "type": "string" } },
          "metadata": { "type": "object", "additionalProperties": true },
          "credential": { "type": "string", "writeOnly": true }
        }
      },
      "EvidenceItem": {
        "type": "object",
        "required": ["externalId", "artifactType", "title", "sourceVersion", "content"],
        "properties": {
          "externalId": { "type": "string" },
          "artifactType": { "type": "string" },
          "title": { "type": "string" },
          "canonicalUrl": { "type": "string", "format": "uri" },
          "sourceVersion": { "type": "string" },
          "content": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true },
          "sourceAcl": { "type": "object", "additionalProperties": true }
        }
      },
      "FindingReview": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "enum": ["CONFIRMED", "DISMISSED", "EDITED"] },
          "reason": { "type": "string" },
          "reasonCode": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/workspaces/{workspaceId}/connections": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "get": { "summary": "List connector status without credentials", "responses": { "200": { "description": "Connections" } } },
      "post": {
        "summary": "Create or update a connector",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConnectionRequest" } } } },
        "responses": { "201": { "description": "Connected" }, "400": { "description": "Invalid provider/request" } }
      }
    },
    "/api/workspaces/{workspaceId}/connections/{connectionId}/sync": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "$ref": "#/components/parameters/ConnectionId" }],
      "post": { "summary": "Run a provider synchronization", "responses": { "202": { "description": "Sync completed/recorded" } } }
    },
    "/api/workspaces/{workspaceId}/connections/{connectionId}/evidence": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "$ref": "#/components/parameters/ConnectionId" }],
      "post": {
        "summary": "Ingest evidence through the provider-neutral boundary",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["artifacts"], "properties": { "artifacts": { "type": "array", "items": { "$ref": "#/components/schemas/EvidenceItem" } }, "reconcile": { "type": "boolean" } } } } } },
        "responses": { "200": { "description": "Sync result" } }
      }
    },
    "/api/workspaces/{workspaceId}/connections/{connectionId}/sync-runs": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "$ref": "#/components/parameters/ConnectionId" }],
      "get": { "summary": "List durable connector sync runs", "responses": { "200": { "description": "Sync runs" } } }
    },
    "/api/connections/{connectionId}": {
      "parameters": [{ "$ref": "#/components/parameters/ConnectionId" }],
      "delete": { "summary": "Revoke credentials and make owned evidence inaccessible", "responses": { "204": { "description": "Revoked" } } }
    },
    "/api/workspaces/{workspaceId}/connections/atlassian/authorize": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "post": { "summary": "Start Atlassian OAuth with selected project/space scope", "responses": { "200": { "description": "Authorization URL and expiring state" } } }
    },
    "/api/integrations/atlassian/oauth/callback": {
      "get": { "security": [], "summary": "Consume the one-time Atlassian OAuth callback", "responses": { "200": { "description": "Connection completed" }, "400": { "description": "Invalid/expired state" } } }
    },
...
    "/api/workspaces/{workspaceId}/changes": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "get": { "summary": "List normalized changes", "parameters": [{ "name": "state", "in": "query", "schema": { "type": "string" } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 100 } }], "responses": { "200": { "description": "Changes" } } }
    },
    "/api/changes/{changeId}": {
      "parameters": [{ "$ref": "#/components/parameters/ChangeId" }],
      "get": { "summary": "Get a change with findings, policies, and feedback", "responses": { "200": { "description": "Change detail" }, "404": { "description": "Not found" } } }
    },
    "/api/changes/{changeId}/reanalyze": {
      "parameters": [{ "$ref": "#/components/parameters/ChangeId" }],
      "post": { "summary": "Queue explicit reanalysis of the current head", "responses": { "202": { "description": "Queued" } } }
    },
    "/api/changes/{changeId}/evaluate-policies": {
      "parameters": [{ "$ref": "#/components/parameters/ChangeId" }],
      "post": { "summary": "Persist current policy evaluations", "responses": { "200": { "description": "Evaluations" } } }
    },
    "/api/changes/{changeId}/policies/dry-run": {
      "parameters": [{ "$ref": "#/components/parameters/ChangeId" }],
      "post": { "summary": "Preview policies without mutation", "responses": { "200": { "description": "Policy previews" } } }
    },
    "/api/changes/{changeId}/findings/{findingId}": {
      "parameters": [{ "$ref": "#/components/parameters/ChangeId" }, { "name": "findingId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "patch": {
        "summary": "Record human feedback on a finding",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FindingReview" } } } },
        "responses": { "200": { "description": "Feedback recorded" } }
      }
    },
    "/api/workspaces/{workspaceId}/evidence": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "get": { "summary": "List evidence artifacts", "parameters": [{ "name": "type", "in": "query", "schema": { "type": "string" } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200 } }], "responses": { "200": { "description": "Artifacts" } } }
    },
...
    "/api/workspaces/{workspaceId}/evidence/{artifactId}": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "name": "artifactId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "get": { "summary": "Get artifact versions and relationships", "responses": { "200": { "description": "Artifact detail" } } }
    },
    "/api/workspaces/{workspaceId}/evidence/import-documents": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "post": { "summary": "Bridge durable document sources into evidence", "responses": { "200": { "description": "Import result" } } }
    },
...
    "/api/workspaces/{workspaceId}/demo/evidence": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "post": { "summary": "Seed a realistic cross-source evidence demo", "responses": { "201": { "description": "Seeded" } } }
    },
...
  }
}